+++ modules/nuxt/playground/nuxt.config.ts
@@ -1,15 +1,4 @@
-import { defineNuxtConfig } from "nuxt/config"
-import chakraModule from "../src/module"
-import type * as NuxtSchema from "@nuxt/schema"
-
+import ChakraUI from "../src/module"
 export default defineNuxtConfig({
-  ssr: true,
-  modules: [chakraModule],
-  chakra: {
-    extendTheme: {
-      colors: {
-        $brand: "#f5f",
-      },
-    },
-  },
+  modules: [ChakraUI],
 })