+++ modules/nuxt/src/index.ts
@@ -1,6 +1,0 @@
-export * from "./module"
-import type * as ThemeTypes from "@chakra-ui/theme"
-
-import { default as ChakraUIModule } from "./module"
-
-export default ChakraUIModule