+++ modules/nuxt/src/runtime/chakra.server.ts
@@ -25,7 +25,9 @@ export function defineNitroPlugin(def: NitroAppPlugin): NitroAppPlugin {
 const CHAKRA_UI_COLOR_MODE_COOKIE_KEY = "chakra-ui-color-mode"
 
 export default defineNitroPlugin((app) => {
-  console.log("chakra-ui-nuxt:server_runtime")
+  if (process.env.NODE_ENV === "development") {
+    console.debug("chakra-ui-nuxt:server_runtime ready")
+  }
   app.hooks.hook("render:html", (html, { event: rawEvent }) => {
     const event = new H3Event(rawEvent.node.req, rawEvent.node.res)
     const parsedCookies = parseCookies(event)