+++ packages/c-focus-lock/src/use-focus-trap.ts
@@ -75,6 +75,7 @@ export function useFocusTrap(
 
   // Restore when `enabled` becomes false
   function restore() {
+    // @ts-expect-error - focusElement is not typed
     focusElement(restoreElement.value)
     restoreElement.value = null
     previousActiveElement.value = null
@@ -126,12 +127,14 @@ export function useFocusTrap(
         if (!contains(containers.value, toElement)) {
           event.preventDefault()
           event.stopPropagation()
+          // @ts-expect-error - focusElement is not typed
           focusElement(previous)
         } else {
           previousActiveElement.value = toElement
           focusElement(toElement)
         }
       } else {
+        // @ts-expect-error - focusElement is not typed
         focusElement(previousActiveElement.value)
       }
     },