+++ packages/core/src/components.ts
@@ -57,6 +57,7 @@ export * from "@chakra-ui/c-scroll-lock"
 // T
 export * from "@chakra-ui/c-theme-provider"
 export * from "@chakra-ui/c-tag"
+export * from "@chakra-ui/c-toast"
 
 // V
 export * from "@chakra-ui/c-visually-hidden"