+++ packages/core/src/index.ts
@@ -1,4 +1,4 @@
-import { computed, Plugin, ref, UnwrapRef } from "vue"
+import { computed, createVNode, Plugin, ref, render, UnwrapRef } from "vue"
 import { theme as defaultTheme, baseTheme, Theme } from "@chakra-ui/theme"
 import { ColorModeRef, setupColorModeContext } from "@chakra-ui/c-color-mode"
 import { toCSSVar, WithCSSVar } from "@chakra-ui/styled-system"
@@ -9,17 +9,19 @@ import {
 } from "@chakra-ui/vue-styled"
 import createCache, { EmotionCache } from "@emotion/cache"
 import internalIcons from "./icon.internals"
-import {
-  extendTheme,
-  extendBaseTheme,
-  ThemeOverride,
-} from "@chakra-ui/theme-utils"
+import type { ThemeOverride } from "@chakra-ui/theme-utils"
 import { MergedIcons, parseIcons } from "./parse-icons"
 import { injectResetStyles, injectThemeGlobalStyles } from "./helpers/css-reset"
 import { mode } from "@chakra-ui/theme-tools"
 import { ChakraPluginOptions } from "./helpers/plugin.types"
-import { Dict } from "@chakra-ui/utils"
-import { localStorageManager, StorageManager } from "@chakra-ui/c-color-mode"
+import { canUseDOM, Dict } from "@chakra-ui/utils"
+import { localStorageManager } from "@chakra-ui/c-color-mode"
+import {
+  ToastContainerId,
+  CToastContainer,
+  ToastContextSymbol,
+  toastContext,
+} from "@chakra-ui/c-toast"
 
 /**
  * 1. Support passing cache options from plugin
@@ -119,13 +121,32 @@ const ChakraUIVuePlugin: Plugin = {
 
     // Set color mode property
     app.config.globalProperties.$mode = mode
+
+    app.provide(ToastContextSymbol, toastContext)
+
+    // Setup toast container component
+    if (canUseDOM()) {
+      const toastContainer =
+        document.getElementById(ToastContainerId) ||
+        document.createElement("div")
+      toastContainer.id = ToastContainerId
+      toastContainer.setAttribute("data-chakra-toast-container", "")
+
+      if (!document.body.contains(toastContainer)) {
+        document.body.insertAdjacentElement("afterend", toastContainer)
+      }
+
+      const vnode = createVNode(CToastContainer)
+      vnode.appContext = app._context
+      render(vnode, toastContainer)
+    }
   },
 }
 
 export type { ChakraPluginOptions }
 export interface ThemeProviderProps extends ThemeOverride {}
 export default ChakraUIVuePlugin
-export { extendTheme, extendBaseTheme }
+export { extendTheme, extendBaseTheme } from "@chakra-ui/theme-utils"
 
 // Export chakra factory function
 export { chakra as chakra }