+++ packages/c-accordion/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @chakra-ui/c-accordion
 
+## 0.0.0-dev-20230215144750
+
+### Patch Changes
+
+- Update build
+- Updated dependencies
+  - @chakra-ui/c-motion@0.0.0-dev-20230215144750
+  - @chakra-ui/vue-system@0.0.0-dev-20230215144750
+  - @chakra-ui/c-icon@0.0.0-dev-20230215144750
+  - @chakra-ui/vue-utils@0.0.0-dev-20230215144750
+  - @chakra-ui/vue-composables@0.0.0-dev-20230215144750
+
 ## 2.0.0-beta.2
 
 ### Minor Changes