+++ packages/c-modal/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @chakra-ui/c-modal
 
+## 0.0.0-dev-20230215144750
+
+### Patch Changes
+
+- Update build
+- Updated dependencies
+  - @chakra-ui/c-focus-lock@0.0.0-dev-20230215144750
+  - @chakra-ui/c-motion@0.0.0-dev-20230215144750
+  - @chakra-ui/vue-accessibilty@0.0.0-dev-20230215144750
+  - @chakra-ui/vue-system@0.0.0-dev-20230215144750
+  - @chakra-ui/c-close-button@0.0.0-dev-20230215144750
+  - @chakra-ui/c-portal@0.0.0-dev-20230215144750
+  - @chakra-ui/c-scroll-lock@0.0.0-dev-20230215144750
+  - @chakra-ui/vue-utils@0.0.0-dev-20230215144750
+  - @chakra-ui/vue-composables@0.0.0-dev-20230215144750
+
 ## 3.0.0-beta.2
 
 ### Minor Changes