+++ packages/styled/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @chakra-ui/vue-styled
 
+## 0.0.0-dev-20230215144750
+
+### Patch Changes
+
+- Update build
+- Updated dependencies
+  - @chakra-ui/vue-utils@0.0.0-dev-20230215144750
+
 ## 1.0.0-beta.2
 
 ### Minor Changes